+++ index.html
@@ -42,7 +42,7 @@
       <div class="footer__text">Я в подвале :(</div>
     </footer>
   </div>
-  <script src="script.js"></script>
+  <script src="script.js" type="module"></script>
 </body>
 
 </html>
